+++ client/src/features/advancedSearch/HelpText.tsx
@@ -5,6 +5,7 @@ import styled from 'styled-components'
 import { useAppSelector } from '../../app/hooks'
 import { IHelpTextKey } from '../../redux/slices/helpTextSlice'
 import InternalLink from '../common/InternalLink'
+import theme from '../../styles/theme'
 
 const StyledP = styled.p`
   text-wrap: initial;
@@ -39,7 +40,12 @@ const HelpText: React.FC = () => {
         top: '0',
       }}
     >
-      <Col>
+      <Col
+        style={{
+          borderBottom: `1px solid ${theme.color.lightGray}`,
+          borderLeft: `1px solid ${theme.color.lightGray}`,
+        }}
+      >
         <h3 style={{ wordWrap: 'break-word' }}>
           Help {helpTextKey !== '' && `for ${helpTextKey}`}
         </h3>