+++ client/src/styles/features/results/ResultsHeader.ts
@@ -6,5 +6,6 @@ const ResultsHeader = styled.h2`
   letter-spacing: 0;
   text-align: left;
   font-weight: 200;
+  margin-right: 0.5em;
 `
 export default ResultsHeader